# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EV Charger Waitlist API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ChargerStatus = typeof ChargerStatus[keyof typeof ChargerStatus];


export const ChargerStatus = {
  available: 'available',
  assigned: 'assigned',
  occupied: 'occupied',
} as const;

export interface Charger {
  id: number;
  name: string;
  status: ChargerStatus;
  /** @nullable */
  currentSessionId?: number | null;
  /** @nullable */
  currentUserName?: string | null;
}

export type QueueEntryStatus = typeof QueueEntryStatus[keyof typeof QueueEntryStatus];


export const QueueEntryStatus = {
  waiting: 'waiting',
  assigned: 'assigned',
  cancelled: 'cancelled',
  forfeited: 'forfeited',
} as const;

export interface QueueEntry {
  id: number;
  userId: string;
  userName: string;
  position: number;
  joinedAt: string;
  status: QueueEntryStatus;
  /** @nullable */
  sessionId?: number | null;
  /** @nullable */
  estimatedWaitMinutes?: number | null;
}

export interface JoinQueueInput {
  note?: string;
}

export interface ClaimDirectInput {
  chargerId: number;
}

export type MyStatusState = typeof MyStatusState[keyof typeof MyStatusState];


export const MyStatusState = {
  not_in_queue: 'not_in_queue',
  waiting: 'waiting',
  assigned: 'assigned',
  checked_in: 'checked_in',
  checked_out: 'checked_out',
} as const;

export type ChargingSessionStatus = typeof ChargingSessionStatus[keyof typeof ChargingSessionStatus];


export const ChargingSessionStatus = {
  assigned: 'assigned',
  claimed: 'claimed',
  checked_in: 'checked_in',
  checked_out: 'checked_out',
  forfeited: 'forfeited',
  expired: 'expired',
} as const;

export interface ChargingSession {
  id: number;
  userId: string;
  userName?: string;
  chargerId: number;
  chargerName: string;
  status: ChargingSessionStatus;
  assignedAt: string;
  claimDeadlineAt?: string;
  /** @nullable */
  claimedAt?: string | null;
  /** @nullable */
  checkedInAt?: string | null;
  /** @nullable */
  checkedOutAt?: string | null;
  /** @nullable */
  minutesRemainingToClaim?: number | null;
}

export interface MyStatus {
  state: MyStatusState;
  queueEntry: QueueEntry | null;
  session: ChargingSession | null;
}

export interface DashboardSummary {
  totalChargers: number;
  availableChargers: number;
  occupiedChargers: number;
  assignedChargers: number;
  queueLength: number;
  averageWaitMinutes: number;
  todayCompletedSessions: number;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ApiError {
  error: string;
}

export type AdminChargerStatus = typeof AdminChargerStatus[keyof typeof AdminChargerStatus];


export const AdminChargerStatus = {
  available: 'available',
  assigned: 'assigned',
  occupied: 'occupied',
} as const;

export interface AdminCharger {
  id: number;
  name: string;
  status: AdminChargerStatus;
  /** @nullable */
  currentSessionId?: number | null;
  /** @nullable */
  currentUserId?: string | null;
  /** @nullable */
  currentUserName?: string | null;
}

export type AdminQueueEntryStatus = typeof AdminQueueEntryStatus[keyof typeof AdminQueueEntryStatus];


export const AdminQueueEntryStatus = {
  waiting: 'waiting',
  assigned: 'assigned',
  cancelled: 'cancelled',
  forfeited: 'forfeited',
} as const;

export interface AdminQueueEntry {
  id: number;
  userId: string;
  userName: string;
  userEmail: string;
  status: AdminQueueEntryStatus;
  joinedAt: string;
  position: number;
  /** @nullable */
  estimatedWaitMinutes?: number | null;
  /** @nullable */
  sessionId?: number | null;
}

export type AdminSessionStatus = typeof AdminSessionStatus[keyof typeof AdminSessionStatus];


export const AdminSessionStatus = {
  assigned: 'assigned',
  claimed: 'claimed',
  checked_in: 'checked_in',
  checked_out: 'checked_out',
  forfeited: 'forfeited',
  expired: 'expired',
} as const;

export interface AdminSession {
  id: number;
  userId: string;
  userName: string;
  chargerId: number;
  chargerName: string;
  status: AdminSessionStatus;
  assignedAt: string;
  claimDeadlineAt: string;
  /** @nullable */
  claimedAt?: string | null;
  /** @nullable */
  checkedInAt?: string | null;
  /** @nullable */
  checkedOutAt?: string | null;
}