# lib/api-zod/src/generated/types/adminCharger.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EV Charger Waitlist API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminChargerStatus } from './adminChargerStatus';

export interface AdminCharger {
  id: number;
  name: string;
  status: AdminChargerStatus;
  /** @nullable */
  currentSessionId?: number | null;
  /** @nullable */
  currentUserId?: string | null;
  /** @nullable */
  currentUserName?: string | null;
}