# lib/api-zod/src/generated/types/adminQueueEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EV Charger Waitlist API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminQueueEntryStatus } from './adminQueueEntryStatus';

export interface AdminQueueEntry {
  id: number;
  userId: string;
  userName: string;
  userEmail: string;
  status: AdminQueueEntryStatus;
  joinedAt: Date;
  position: number;
  /** @nullable */
  estimatedWaitMinutes?: number | null;
  /** @nullable */
  sessionId?: number | null;
}