# lib/api-zod/src/generated/types/adminSession.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * EV Charger Waitlist API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminSessionStatus } from './adminSessionStatus';

export interface AdminSession {
  id: number;
  userId: string;
  userName: string;
  chargerId: number;
  chargerName: string;
  status: AdminSessionStatus;
  assignedAt: Date;
  claimDeadlineAt: Date;
  /** @nullable */
  claimedAt?: Date | null;
  /** @nullable */
  checkedInAt?: Date | null;
  /** @nullable */
  checkedOutAt?: Date | null;
}